/test-s3 endpoint to verify S3 credentials
