+++ apps/mobile-server/src/throwaway/throwaway.controller.ts
@@ -7,25 +7,6 @@ import { Request } from 'express';
 export class ThrowawayController {
   constructor(@Inject('REDIS_CLIENT') private readonly redisClient: Redis) {}
 
-  @Get('test')
-  async testAction(@Session() session: Record<string, string>) {
-    const options = {
-      region: 'eu-west-1',
-      credentials: JSON.parse(session.awsCredentials),
-    };
-
-    const s3Client = new S3Client(options);
-
-    const listCommand = new ListObjectsV2Command({
-      Bucket: 'jvalanen-diory-test3',
-      Prefix: 'room/',
-    });
-
-    const list = await s3Client.send(listCommand);
-
-    return JSON.stringify(list);
-  }
-
   @Get('rooms')
   async roomsAction(@Session() session: Record<string, string>) {
     const response = await this.redisClient.keys(`${session.userId}-rooms-*`);
